Fix misspelled methods and cleanup job model/tests
Changing formated_* methods to be formatted_*. Just a minor spelling
correction, which led to updating the company_spec.

Updating company_spec led to to refactoring company_spec, and part of
that was to change company.*? methods to company.has_*? methods. This
makes the tests and code read a little better, e.g.:

company.phone1?.should be_false
=>
company.should_not have_phone1

and

<% if company.website? %>
=>
<% if company.has_website? %>

Also converting tabs => spaces if I happen to be working in a file.